fix: Resolve clippy warnings with idiomatic Rust
Fixed:
- session.rs: Use iterator for secret initialization
- nonce.rs: Use iterator for random_base and nonce XOR

All 8 tests pass. Warnings reduced from 7 to 5.

+++ omni-daemon/src/tun.rs
@@ -55,7 +55,7 @@ impl VirtualInterface {
             .tun_name(&config.name)
             .address(config.address)
             .netmask(config.netmask)
-            .mtu(config.mtu as u16)
+            .mtu(config.mtu)
             .up();
         
         // Platform-specific configuration
